Move limitation cases to separate test case
