+++ tests/integration/Core/Framework/DataAbstractionLayer/Search/JoinFilterTest.php
@@ -32,6 +32,8 @@ use Shopware\Core\Test\Stub\Framework\IdsCollection;
 
 /**
  * @internal
+ *
+ * @see MultiJoinFilterLimitationTest for edge cases and limitations of multi join filters
  */
 class JoinFilterTest extends TestCase
 {
@@ -471,64 +473,6 @@ class JoinFilterTest extends TestCase
     }
 
     #[Depends('testIndexing')]
-    public function testOneToManyWithSortWithMultipleJoinGroups(IdsCollection $ids): void
-    {
-        $criteria = new Criteria($ids->prefixed('product-'));
-        $criteria->addFilter(
-            new OrFilter([
-                new AndFilter([
-                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-2')),
-                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
-                ]),
-                new AndFilter([
-                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-1')),
-                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
-                ]),
-            ])
-        );
-        $criteria->addSorting(new FieldSorting('product.prices.price'));
-
-        $result = static::getContainer()->get('product.repository')
-            ->searchIds($criteria, Context::createDefaultContext());
-
-        static::assertSame(2, $result->getTotal());
-        // note that this is the same order then below, because apparently it uses both rule-1 price and rule-2 price
-        // for sorting, therefore product-1 comes first in both cases, as it has same higher and lower price
-        // however note that by the join group the lower rule-1 price should be filtered out
-        static::assertSame($ids->get('product-1'), $result->getIds()[0]);
-        static::assertSame($ids->get('product-2'), $result->getIds()[1]);
-    }
-
-    #[Depends('testIndexing')]
-    public function testOneToManyWithSortWithMultipleJoinGroupsDesc(IdsCollection $ids): void
-    {
-        $criteria = new Criteria($ids->prefixed('product-'));
-        $criteria->addFilter(
-            new OrFilter([
-                new AndFilter([
-                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-1')),
-                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
-                ]),
-                new AndFilter([
-                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-2')),
-                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
-                ]),
-            ])
-        );
-        $criteria->addSorting(new FieldSorting('product.prices.price', FieldSorting::DESCENDING));
-
-        $result = static::getContainer()->get('product.repository')
-            ->searchIds($criteria, Context::createDefaultContext());
-
-        static::assertSame(2, $result->getTotal());
-        // note that this is the same order then above, because apparently it uses both rule-1 price and rule-2 price
-        // for sorting, therefore product-1 comes first in both cases, as it has same higher and lower price
-        // however note that by the join group the lower rule-1 price should be filtered out
-        static::assertSame($ids->get('product-1'), $result->getIds()[0]);
-        static::assertSame($ids->get('product-2'), $result->getIds()[1]);
-    }
-
-    #[Depends('testIndexing')]
     public function testOneToManyWithGrouping(IdsCollection $ids): void
     {
         $criteria = new Criteria($ids->prefixed('product-'));
@@ -545,31 +489,6 @@ class JoinFilterTest extends TestCase
 
         static::assertSame(1, $result->getTotal());
         static::assertSame($ids->get('product-1'), $result->getIds()[0]);
-    }
-
-    #[Depends('testIndexing')]
-    public function testOneToManyWithMultipleJoinGroupsAndGroupingIsNotSupported(IdsCollection $ids): void
-    {
-        $criteria = new Criteria($ids->prefixed('product-'));
-        $criteria->addFilter(
-            new OrFilter([
-                new AndFilter([
-                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-1')),
-                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
-                ]),
-                new AndFilter([
-                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-2')),
-                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
-                ]),
-            ])
-        );
-        $criteria->addGroupField(new FieldGrouping('product.prices.ruleId'));
-
-        // we get an internal DBAL field mapping exception here,
-        // which is not optimal but at least indicates that this is not supported
-        static::expectException(\Throwable::class);
-        static::getContainer()->get('product.repository')
-            ->searchIds($criteria, Context::createDefaultContext());
     }
 
     #[Depends('testIndexing')]
@@ -610,87 +529,6 @@ class JoinFilterTest extends TestCase
         static::assertTrue($result->has($ids->get('category-1')));
         static::assertTrue($result->has($ids->get('category-2')));
         static::assertFalse($result->has($ids->get('category-3')));
-    }
-
-    #[Depends('testIndexing')]
-    public function testManyToOneWithSort(IdsCollection $ids): void
-    {
-        $criteria = new Criteria($ids->prefixed('category-'));
-
-        $criteria->addFilter(
-            new OrFilter([
-                new AndFilter([
-                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-1')),
-                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-1'),
-                ]),
-                new AndFilter([
-                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-2')),
-                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-2'),
-                ]),
-            ])
-        );
-        $criteria->addSorting(new FieldSorting('category.products.manufacturer.name'));
-
-        $result = static::getContainer()->get('category.repository')
-            ->searchIds($criteria, Context::createDefaultContext());
-
-        static::assertSame(3, $result->getTotal());
-        static::assertSame($ids->get('category-1'), $result->getIds()[0]);
-        static::assertSame($ids->get('category-2'), $result->getIds()[1]);
-        static::assertSame($ids->get('category-3'), $result->getIds()[2]);
-    }
-
-    #[Depends('testIndexing')]
-    public function testManyToOneWithSortDesc(IdsCollection $ids): void
-    {
-        $criteria = new Criteria($ids->prefixed('category-'));
-
-        $criteria->addFilter(
-            new OrFilter([
-                new AndFilter([
-                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-1')),
-                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-1'),
-                ]),
-                new AndFilter([
-                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-2')),
-                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-2'),
-                ]),
-            ])
-        );
-        $criteria->addSorting(new FieldSorting('category.products.manufacturer.name', FieldSorting::DESCENDING));
-
-        $result = static::getContainer()->get('category.repository')
-            ->searchIds($criteria, Context::createDefaultContext());
-
-        static::assertSame(3, $result->getTotal());
-        static::assertSame($ids->get('category-1'), $result->getIds()[0]); // manufacturer-2 matches as well
-        static::assertSame($ids->get('category-3'), $result->getIds()[1]); // manufacturer-2
-        static::assertSame($ids->get('category-2'), $result->getIds()[2]); // manufacturer-1
-    }
-
-    #[Depends('testIndexing')]
-    public function testManyToOneWithMultipleJoinGroupsAndGroupingIsNotSupported(IdsCollection $ids): void
-    {
-        $criteria = new Criteria($ids->prefixed('category-'));
-        $criteria->addFilter(
-            new OrFilter([
-                new AndFilter([
-                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-1')),
-                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-1'),
-                ]),
-                new AndFilter([
-                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-2')),
-                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-2'),
-                ]),
-            ])
-        );
-        $criteria->addGroupField(new FieldGrouping('category.products.manufacturer.name'));
-
-        // we get an internal DBAL field mapping exception here,
-        // which is not optimal but at least indicates that this is not supported
-        static::expectException(\Throwable::class);
-        static::getContainer()->get('category.repository')
-            ->searchIds($criteria, Context::createDefaultContext());
     }
 
     #[Depends('testIndexing')]
@@ -735,83 +573,6 @@ class JoinFilterTest extends TestCase
         static::assertSame(2, $result->getTotal());
         static::assertTrue($result->has($ids->get('product-1')));
         static::assertTrue($result->has($ids->get('product-2')));
-    }
-
-    #[Depends('testIndexing')]
-    public function testManyToManyWithSort(IdsCollection $ids): void
-    {
-        $criteria = new Criteria($ids->prefixed('product-'));
-        $criteria->addFilter(
-            new OrFilter([
-                new AndFilter([
-                    new EqualsFilter('product.properties.id', $ids->get('yellow')),
-                    new EqualsFilter('product.properties.name', 'yellow'),
-                ]),
-                new AndFilter([
-                    new EqualsFilter('product.properties.id', $ids->get('S')),
-                    new EqualsFilter('product.properties.name', 'S'),
-                ]),
-            ])
-        );
-        $criteria->addSorting(new FieldSorting('product.properties.name'));
-
-        $result = static::getContainer()->get('product.repository')
-            ->searchIds($criteria, Context::createDefaultContext());
-
-        static::assertSame(2, $result->getTotal());
-        static::assertSame($ids->get('product-1'), $result->getIds()[0]);
-        static::assertSame($ids->get('product-2'), $result->getIds()[1]);
-    }
-
-    #[Depends('testIndexing')]
-    public function testManyToManyWithSortDesc(IdsCollection $ids): void
-    {
-        $criteria = new Criteria($ids->prefixed('product-'));
-        $criteria->addFilter(
-            new OrFilter([
-                new AndFilter([
-                    new EqualsFilter('product.properties.id', $ids->get('yellow')),
-                    new EqualsFilter('product.properties.name', 'yellow'),
-                ]),
-                new AndFilter([
-                    new EqualsFilter('product.properties.id', $ids->get('S')),
-                    new EqualsFilter('product.properties.name', 'S'),
-                ]),
-            ])
-        );
-        $criteria->addSorting(new FieldSorting('product.properties.name', FieldSorting::DESCENDING));
-
-        $result = static::getContainer()->get('product.repository')
-            ->searchIds($criteria, Context::createDefaultContext());
-
-        static::assertSame(2, $result->getTotal());
-        static::assertSame($ids->get('product-1'), $result->getIds()[0]);
-        static::assertSame($ids->get('product-2'), $result->getIds()[1]);
-    }
-
-    #[Depends('testIndexing')]
-    public function testManyToManyWithGroup(IdsCollection $ids): void
-    {
-        $criteria = new Criteria($ids->prefixed('product-'));
-        $criteria->addFilter(
-            new OrFilter([
-                new AndFilter([
-                    new EqualsFilter('product.properties.id', $ids->get('yellow')),
-                    new EqualsFilter('product.properties.name', 'yellow'),
-                ]),
-                new AndFilter([
-                    new EqualsFilter('product.properties.id', $ids->get('S')),
-                    new EqualsFilter('product.properties.name', 'S'),
-                ]),
-            ])
-        );
-        $criteria->addGroupField(new FieldGrouping('product.properties.name'));
-
-        // we get an internal DBAL field mapping exception here,
-        // which is not optimal but at least indicates that this is not supported
-        static::expectException(\Throwable::class);
-        static::getContainer()->get('product.repository')
-            ->searchIds($criteria, Context::createDefaultContext());
     }
 
     #[Depends('testIndexing')]
