+++ tests/integration/Core/Framework/DataAbstractionLayer/Search/MultiJoinFilterLimitationTest.php
@@ -1,0 +1,368 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Tests\Integration\Core\Framework\DataAbstractionLayer\Search;
+
+use Doctrine\DBAL\Connection;
+use PHPUnit\Framework\Attributes\AfterClass;
+use PHPUnit\Framework\Attributes\BeforeClass;
+use PHPUnit\Framework\Attributes\Depends;
+use PHPUnit\Framework\TestCase;
+use Shopware\Core\Content\Test\Product\ProductBuilder;
+use Shopware\Core\Framework\Context;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\AndFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\OrFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\RangeFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Grouping\FieldGrouping;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Sorting\FieldSorting;
+use Shopware\Core\Framework\Test\TestCaseBase\KernelLifecycleManager;
+use Shopware\Core\Framework\Test\TestCaseBase\KernelTestBehaviour;
+use Shopware\Core\Test\Stub\Framework\IdsCollection;
+
+/**
+ * @internal
+ *
+ * This test case covers known limitations when using multiple join groups
+ * Due to conceptual reasons multi join groups won't do a "real join" to the filtered association,
+ * this means all other DAL features (e.g. sorting, grouping) won't work as expected in combination with multi join groups.
+ * Sorting is based on a unfiltered join (meaning all associated entities are considered for sorting, not just the filtered once).
+ * Grouping is only supported in conjunction with sorting in those cases, and would then also operate on the unfiltered join.
+ *
+ * The behaviour documented in test explicitly is not considered part of the public API and therefore might be fixed in future versions.
+ * The purpose of this test is mainly to make the current limitations explicit and to avoid accidental changes to the current behaviour.
+ *
+ * @see JoinFilterTest for the tests for all valid cases that are part of the public API.
+ */
+class MultiJoinFilterLimitationTest extends TestCase
+{
+    use KernelTestBehaviour;
+
+    #[BeforeClass]
+    public static function startTransactionBefore(): void
+    {
+        $connection = KernelLifecycleManager::getKernel()
+            ->getContainer()
+            ->get(Connection::class);
+
+        $connection->beginTransaction();
+    }
+
+    #[AfterClass]
+    public static function stopTransactionAfter(): void
+    {
+        $connection = KernelLifecycleManager::getKernel()
+            ->getContainer()
+            ->get(Connection::class);
+
+        $connection->rollBack();
+    }
+
+    /**
+     * @return IdsCollection
+     */
+    public function testIndexing()
+    {
+        $ids = new IdsCollection();
+
+        $products = [
+            (new ProductBuilder($ids, 'product-1', 10, 'tax'))
+                ->price(15, 10)
+                ->manufacturer('manufacturer-1')
+                ->property('red', 'color')
+                ->property('yellow', 'color')
+                ->property('XL', 'size')
+                ->property('L', 'size')
+                ->category('category-1')
+                ->category('category-2')
+                ->prices('rule-1', 100)
+                ->prices('rule-2', 150)
+                ->build(),
+
+            (new ProductBuilder($ids, 'product-1-variant', 10, 'tax'))
+                ->parent('product-1')
+                ->build(),
+
+            (new ProductBuilder($ids, 'product-2', 3, 'tax'))
+                ->price(15, 10)
+                ->manufacturer('manufacturer-2')
+                ->property('red', 'color')
+                ->property('S', 'size')
+                ->category('category-1')
+                ->category('category-3')
+                ->prices('rule-1', 150)
+                ->build(),
+
+            (new ProductBuilder($ids, 'product-3', 3, 'tax'))
+                ->price(15, 10)
+                ->category('category-4')
+                ->build(),
+        ];
+
+        static::getContainer()->get('product.repository')
+            ->create($products, Context::createDefaultContext());
+
+        $userId = static::getContainer()->get(Connection::class)
+            ->fetchOne('SELECT LOWER(HEX(id)) FROM `user`');
+
+        $ids->set('user-id', $userId);
+
+        $media = [
+            ['id' => $ids->create('with-avatar')],
+            ['id' => $ids->create('without-avatar')],
+        ];
+
+        static::getContainer()->get('media.repository')
+            ->create($media, Context::createDefaultContext());
+
+        $avatar = [
+            'id' => $userId,
+            'avatarId' => $ids->get('with-avatar'),
+        ];
+
+        static::getContainer()->get('user.repository')
+            ->update([$avatar], Context::createDefaultContext());
+
+        $result = static::getContainer()->get('product.repository')
+            ->searchIds(new Criteria($ids->prefixed('product-')), Context::createDefaultContext());
+
+        static::assertSame(\count($products), $result->getTotal());
+
+        return $ids;
+    }
+
+    #[Depends('testIndexing')]
+    public function testOneToManyWithSortWithMultipleJoinGroups(IdsCollection $ids): void
+    {
+        $criteria = new Criteria($ids->prefixed('product-'));
+        $criteria->addFilter(
+            new OrFilter([
+                new AndFilter([
+                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-2')),
+                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
+                ]),
+                new AndFilter([
+                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-1')),
+                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
+                ]),
+            ])
+        );
+        $criteria->addSorting(new FieldSorting('product.prices.price'));
+
+        $result = static::getContainer()->get('product.repository')
+            ->searchIds($criteria, Context::createDefaultContext());
+
+        static::assertSame(2, $result->getTotal());
+        // note that this is the same order then below, because apparently it uses both rule-1 price and rule-2 price
+        // for sorting, therefore product-1 comes first in both cases, as it has same higher and lower price
+        // however note that by the join group the lower rule-1 price should be filtered out
+        static::assertSame($ids->get('product-1'), $result->getIds()[0]);
+        static::assertSame($ids->get('product-2'), $result->getIds()[1]);
+    }
+
+    #[Depends('testIndexing')]
+    public function testOneToManyWithSortWithMultipleJoinGroupsDesc(IdsCollection $ids): void
+    {
+        $criteria = new Criteria($ids->prefixed('product-'));
+        $criteria->addFilter(
+            new OrFilter([
+                new AndFilter([
+                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-1')),
+                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
+                ]),
+                new AndFilter([
+                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-2')),
+                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
+                ]),
+            ])
+        );
+        $criteria->addSorting(new FieldSorting('product.prices.price', FieldSorting::DESCENDING));
+
+        $result = static::getContainer()->get('product.repository')
+            ->searchIds($criteria, Context::createDefaultContext());
+
+        static::assertSame(2, $result->getTotal());
+        // note that this is the same order then above, because apparently it uses both rule-1 price and rule-2 price
+        // for sorting, therefore product-1 comes first in both cases, as it has same higher and lower price
+        // however note that by the join group the lower rule-1 price should be filtered out
+        static::assertSame($ids->get('product-1'), $result->getIds()[0]);
+        static::assertSame($ids->get('product-2'), $result->getIds()[1]);
+    }
+
+    #[Depends('testIndexing')]
+    public function testOneToManyWithMultipleJoinGroupsAndGroupingIsNotSupported(IdsCollection $ids): void
+    {
+        $criteria = new Criteria($ids->prefixed('product-'));
+        $criteria->addFilter(
+            new OrFilter([
+                new AndFilter([
+                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-1')),
+                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
+                ]),
+                new AndFilter([
+                    new EqualsFilter('product.prices.ruleId', $ids->get('rule-2')),
+                    new RangeFilter('product.prices.price', [RangeFilter::GTE => 150]),
+                ]),
+            ])
+        );
+        $criteria->addGroupField(new FieldGrouping('product.prices.ruleId'));
+
+        static::expectException(\Throwable::class);
+        static::getContainer()->get('product.repository')
+            ->searchIds($criteria, Context::createDefaultContext());
+    }
+
+    #[Depends('testIndexing')]
+    public function testManyToOneWithSort(IdsCollection $ids): void
+    {
+        $criteria = new Criteria($ids->prefixed('category-'));
+
+        $criteria->addFilter(
+            new OrFilter([
+                new AndFilter([
+                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-1')),
+                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-1'),
+                ]),
+                new AndFilter([
+                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-2')),
+                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-2'),
+                ]),
+            ])
+        );
+        $criteria->addSorting(new FieldSorting('category.products.manufacturer.name'));
+
+        $result = static::getContainer()->get('category.repository')
+            ->searchIds($criteria, Context::createDefaultContext());
+
+        static::assertSame(3, $result->getTotal());
+        static::assertSame($ids->get('category-1'), $result->getIds()[0]);
+        static::assertSame($ids->get('category-2'), $result->getIds()[1]);
+        static::assertSame($ids->get('category-3'), $result->getIds()[2]);
+    }
+
+    #[Depends('testIndexing')]
+    public function testManyToOneWithSortDesc(IdsCollection $ids): void
+    {
+        $criteria = new Criteria($ids->prefixed('category-'));
+
+        $criteria->addFilter(
+            new OrFilter([
+                new AndFilter([
+                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-1')),
+                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-1'),
+                ]),
+                new AndFilter([
+                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-2')),
+                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-2'),
+                ]),
+            ])
+        );
+        $criteria->addSorting(new FieldSorting('category.products.manufacturer.name', FieldSorting::DESCENDING));
+
+        $result = static::getContainer()->get('category.repository')
+            ->searchIds($criteria, Context::createDefaultContext());
+
+        static::assertSame(3, $result->getTotal());
+        static::assertSame($ids->get('category-1'), $result->getIds()[0]); // manufacturer-2 matches as well
+        static::assertSame($ids->get('category-3'), $result->getIds()[1]); // manufacturer-2
+        static::assertSame($ids->get('category-2'), $result->getIds()[2]); // manufacturer-1
+    }
+
+    #[Depends('testIndexing')]
+    public function testManyToOneWithMultipleJoinGroupsAndGroupingIsNotSupported(IdsCollection $ids): void
+    {
+        $criteria = new Criteria($ids->prefixed('category-'));
+        $criteria->addFilter(
+            new OrFilter([
+                new AndFilter([
+                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-1')),
+                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-1'),
+                ]),
+                new AndFilter([
+                    new EqualsFilter('category.products.manufacturer.id', $ids->get('manufacturer-2')),
+                    new EqualsFilter('category.products.manufacturer.name', 'manufacturer-2'),
+                ]),
+            ])
+        );
+        $criteria->addGroupField(new FieldGrouping('category.products.manufacturer.name'));
+
+        static::expectException(\Throwable::class);
+        static::getContainer()->get('category.repository')
+            ->searchIds($criteria, Context::createDefaultContext());
+    }
+
+    #[Depends('testIndexing')]
+    public function testManyToManyWithSort(IdsCollection $ids): void
+    {
+        $criteria = new Criteria($ids->prefixed('product-'));
+        $criteria->addFilter(
+            new OrFilter([
+                new AndFilter([
+                    new EqualsFilter('product.properties.id', $ids->get('yellow')),
+                    new EqualsFilter('product.properties.name', 'yellow'),
+                ]),
+                new AndFilter([
+                    new EqualsFilter('product.properties.id', $ids->get('S')),
+                    new EqualsFilter('product.properties.name', 'S'),
+                ]),
+            ])
+        );
+        $criteria->addSorting(new FieldSorting('product.properties.name'));
+
+        $result = static::getContainer()->get('product.repository')
+            ->searchIds($criteria, Context::createDefaultContext());
+
+        static::assertSame(2, $result->getTotal());
+        static::assertSame($ids->get('product-1'), $result->getIds()[0]);
+        static::assertSame($ids->get('product-2'), $result->getIds()[1]);
+    }
+
+    #[Depends('testIndexing')]
+    public function testManyToManyWithSortDesc(IdsCollection $ids): void
+    {
+        $criteria = new Criteria($ids->prefixed('product-'));
+        $criteria->addFilter(
+            new OrFilter([
+                new AndFilter([
+                    new EqualsFilter('product.properties.id', $ids->get('yellow')),
+                    new EqualsFilter('product.properties.name', 'yellow'),
+                ]),
+                new AndFilter([
+                    new EqualsFilter('product.properties.id', $ids->get('S')),
+                    new EqualsFilter('product.properties.name', 'S'),
+                ]),
+            ])
+        );
+        $criteria->addSorting(new FieldSorting('product.properties.name', FieldSorting::DESCENDING));
+
+        $result = static::getContainer()->get('product.repository')
+            ->searchIds($criteria, Context::createDefaultContext());
+
+        static::assertSame(2, $result->getTotal());
+        static::assertSame($ids->get('product-1'), $result->getIds()[0]);
+        static::assertSame($ids->get('product-2'), $result->getIds()[1]);
+    }
+
+    #[Depends('testIndexing')]
+    public function testManyToManyWithGroup(IdsCollection $ids): void
+    {
+        $criteria = new Criteria($ids->prefixed('product-'));
+        $criteria->addFilter(
+            new OrFilter([
+                new AndFilter([
+                    new EqualsFilter('product.properties.id', $ids->get('yellow')),
+                    new EqualsFilter('product.properties.name', 'yellow'),
+                ]),
+                new AndFilter([
+                    new EqualsFilter('product.properties.id', $ids->get('S')),
+                    new EqualsFilter('product.properties.name', 'S'),
+                ]),
+            ])
+        );
+        $criteria->addGroupField(new FieldGrouping('product.properties.name'));
+        
+        static::expectException(\Throwable::class);
+        static::getContainer()->get('product.repository')
+            ->searchIds($criteria, Context::createDefaultContext());
+    }
+}
